+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.17.0-0.7.1 (2025-05-28)
 
+* [PLT-2244] Disable setting CRIVolume by default
 * [PLT-2099] Fix coredns PDB specification
 * [PLT-2131] Improve workers checks during cloud-provisioner upgrade to avoid timeouts
 * [PLT-2098] Improve kubernetes version checks during cloud-provisioner-upgrade